Fixes stateFullNameFromAbbreviation method
The method was incorrect.  Line 31 was attempting to search for the located value as a key for the dictionary.

Ex.  [@"HI" stateFullNameFromAbbreviation]:

Line 30: The loop finds "HI" in the values for the dictionary
Line 31: [self.stateAbbreviationsMap objectForKey:@"HI"] returns nil because the key is "Hawaii" and the value is "HI"